Split 1.17 module into 1.17-common, 1.17, and 1.17.1
