+++ commandapi-nms/commandapi-1.17-common/.gitignore
@@ -1,0 +1,2 @@
+/.apt_generated/
+/.apt_generated_tests/
