+++ commandapi-nms/commandapi-1.17-common/pom.xml
@@ -25,22 +25,38 @@
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<artifactId>commandapi-1.17</artifactId>
+	<artifactId>commandapi-1.17-common</artifactId>
 
 	<properties>
-		<minecraft.mappings>1.17-R0.1-SNAPSHOT</minecraft.mappings>
+		<minecraft.mappings>1.17.1-R0.1-SNAPSHOT</minecraft.mappings>
 		<spigot.version>1.17.1-R0.1-SNAPSHOT</spigot.version>
 	</properties>
 
+	<repositories>
+		<repository>
+			<id>minecraft-libraries</id>
+			<url>https://libraries.minecraft.net</url>
+		</repository>
+		<repository>
+			<id>codemc-repo</id>
+			<url>https://repo.codemc.io/repository/nms/</url>
+		</repository>
+		<repository>
+			<id>papermc</id>
+			<url>https://papermc.io/repo/repository/maven-public/</url>
+		</repository>
+	</repositories>
+
 	<dependencies>
+		<!-- Spigot testing (for reflection checks) -->
 		<dependency>
-			<groupId>dev.jorel</groupId>
-			<artifactId>commandapi-1.17-common</artifactId>
-			<version>${project.version}</version>
+			<groupId>org.spigotmc</groupId>
+			<artifactId>spigot</artifactId>
+			<version>${spigot.version}</version>
+			<scope>test</scope>
 		</dependency>
 
 		<!-- Spigot compilation -->
-		<!-- Include both mojang mapped and unmapped spigot so 1_17 common can be accessed before and after it is remapped -->
 		<dependency>
 			<groupId>org.spigotmc</groupId>
 			<artifactId>spigot</artifactId>
@@ -48,11 +64,28 @@
 			<classifier>remapped-mojang</classifier>
 			<scope>provided</scope>
 		</dependency>
+
+		<!-- Libraries -->
 		<dependency>
-			<groupId>org.spigotmc</groupId>
-			<artifactId>spigot</artifactId>
-			<version>${spigot.version}</version>
+			<groupId>io.papermc.paper</groupId>
+			<artifactId>paper-api</artifactId>
+			<version>${paper.version}</version>
 			<scope>provided</scope>
+		</dependency>
+
+		<!-- CommandAPI core -->
+		<dependency>
+			<groupId>dev.jorel</groupId>
+			<artifactId>commandapi-core</artifactId>
+			<version>${project.version}</version>
+			<scope>compile</scope>
+		</dependency>
+		
+		<dependency>
+			<groupId>dev.jorel</groupId>
+			<artifactId>commandapi-nms-common</artifactId>
+			<version>${project.version}</version>
+			<scope>compile</scope>
 		</dependency>
 	</dependencies>
