+++ commandapi-nms/commandapi-1.17-common/src/test/java/SafeReflect.java
