+++ commandapi-nms/commandapi-1.17.1/src/main/java/dev/jorel/commandapi/nms/NMS_1_17_R1.java
@@ -1,0 +1,105 @@
+/*******************************************************************************
+ * Copyright 2018, 2021 Jorel Ali (Skepter) - MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *******************************************************************************/
+package dev.jorel.commandapi.nms;
+
+import dev.jorel.commandapi.CommandAPI;
+import dev.jorel.commandapi.CommandAPIHandler;
+import dev.jorel.commandapi.preprocessor.NMSMeta;
+import net.minecraft.Util;
+import net.minecraft.server.ServerFunctionLibrary;
+import net.minecraft.server.ServerResources;
+import net.minecraft.server.packs.resources.ReloadableResourceManager;
+import org.bukkit.Bukkit;
+import org.bukkit.Keyed;
+import org.bukkit.inventory.Recipe;
+
+import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.util.Iterator;
+import java.util.concurrent.CompletableFuture;
+
+/**
+ * NMS implementation for Minecraft 1.17.1
+ */
+@NMSMeta(compatibleWith = { "1.17.1"})
+public class NMS_1_17_R1 extends NMS_1_17_Common {
+	@Override
+	public void reloadDataPacks() {
+		CommandAPI.logNormal("Reloading datapacks...");
+
+		// Get previously declared recipes to be re-registered later
+		Iterator<Recipe> recipes = Bukkit.recipeIterator();
+
+		// Update the commandDispatcher with the current server's commandDispatcher
+		ServerResources serverResources = MINECRAFT_SERVER.resources;
+		serverResources.commands = MINECRAFT_SERVER.getCommands();
+
+		// Update the ServerFunctionLibrary's command dispatcher with the new one
+		try {
+			CommandAPIHandler.getInstance().getField(ServerFunctionLibrary.class, "i")
+				.set(serverResources.getFunctionLibrary(), getBrigadierDispatcher());
+		} catch (ReflectiveOperationException e) {
+			e.printStackTrace();
+		}
+
+		// Construct the new CompletableFuture that now uses our updated serverResources
+		CompletableFuture<?> unitCompletableFuture = ((ReloadableResourceManager) serverResources.getResourceManager())
+			.reload(Util.backgroundExecutor(), Runnable::run,
+				MINECRAFT_SERVER.getPackRepository().openAllSelected(),
+				CompletableFuture.completedFuture(null));
+		CompletableFuture<ServerResources> completablefuture = unitCompletableFuture
+			.whenComplete((Object u, Throwable t) -> {
+				if (t != null) {
+					serverResources.close();
+				}
+			}).thenApply((Object u) -> serverResources);
+
+		// Run the completableFuture and bind tags
+		try {
+			completablefuture.get().updateGlobals();
+
+			// Register recipes again because reloading datapacks removes all non-vanilla
+			// recipes
+			Recipe recipe;
+			while (recipes.hasNext()) {
+				recipe = recipes.next();
+				try {
+					Bukkit.addRecipe(recipe);
+					if (recipe instanceof Keyed keyedRecipe) {
+						CommandAPI.logInfo("Re-registering recipe: " + keyedRecipe.getKey());
+					}
+				} catch (Exception e) {
+					continue; // Can't re-register registered recipes. Not an error.
+				}
+			}
+
+			CommandAPI.logNormal("Finished reloading datapacks");
+		} catch (Exception e) {
+			StringWriter stringWriter = new StringWriter();
+			PrintWriter printWriter = new PrintWriter(stringWriter);
+			e.printStackTrace(printWriter);
+
+			CommandAPI.logError(
+				"Failed to load datapacks, can't proceed with normal server load procedure. Try fixing your datapacks?\n"
+					+ stringWriter.toString());
+		}
+	}
+}
