+++ commandapi-nms/commandapi-nms-dependency/pom.xml
@@ -57,6 +57,11 @@
 		</dependency>
 		<dependency>
 			<groupId>dev.jorel</groupId>
+			<artifactId>commandapi-1.17.1</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>dev.jorel</groupId>
 			<artifactId>commandapi-1.16.5</artifactId>
 			<version>${project.version}</version>
 		</dependency>
